requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
rdkit-pypi>=2022.9.1
scikit-learn>=1.1.0
chembl-webresource-client>=0.10.0
plotly>=5.0.0
scipy>=1.9.0